lara/common: move Lara_Animate to TRX
This makes the Lara_Animate function common in TRX.
