+++ src/libtrx/include/libtrx/game/lara/common.h
@@ -5,3 +5,4 @@
 
 LARA_INFO *Lara_GetLaraInfo(void);
 ITEM *Lara_GetItem(void);
+void Lara_Animate(ITEM *item);
