+++ src/tr1/game/lara/common.h
@@ -4,6 +4,8 @@
 
 #include "global/types.h"
 
+#include <libtrx/game/lara/common.h>
+
 #include <stdint.h>
 
 void Lara_Control(void);
@@ -12,7 +14,6 @@ ITEM *Lara_GetDeathCameraTarget(void);
 void Lara_SetDeathCameraTarget(int16_t item_num);
 
 void Lara_ControlExtra(int16_t item_num);
-void Lara_Animate(ITEM *item);
 void Lara_AnimateUntil(ITEM *lara_item, int32_t goal);
 
 void Lara_Initialise(int32_t level_num);
