+++ src/tr2/game/lara/control.h
@@ -2,6 +2,8 @@
 
 #include "global/types.h"
 
+#include <libtrx/game/lara/common.h>
+
 void Lara_HandleAboveWater(ITEM *item, COLL_INFO *coll);
 
 void Lara_HandleSurface(ITEM *item, COLL_INFO *coll);
@@ -10,8 +12,6 @@ void Lara_HandleUnderwater(ITEM *item, COLL_INFO *coll);
 
 void Lara_Control(int16_t item_num);
 void Lara_ControlExtra(int16_t item_num);
-
-void Lara_Animate(ITEM *item);
 
 void Lara_UseItem(GAME_OBJECT_ID object_id);
